+++ components-api/src/main/scala/pl/touk/nussknacker/engine/api/component/AdditionalUIConfigProvider.scala
@@ -2,11 +2,7 @@ package pl.touk.nussknacker.engine.api.component
 
 import io.circe.generic.JsonCodec
 import pl.touk.nussknacker.engine.api.definition.FixedExpressionValue
-import pl.touk.nussknacker.engine.api.parameter.{
-  ParameterName,
-  ParameterValueCompileTimeValidation,
-  ParameterValueInput
-}
+import pl.touk.nussknacker.engine.api.parameter.{ParameterName, ParameterValueInput}
 
 /**
  * Trait allowing the provision of UI configuration for components and scenario properties.
@@ -39,6 +35,5 @@ case class ParameterAdditionalUIConfig(
     required: Boolean,
     initialValue: Option[FixedExpressionValue],
     hintText: Option[String],
-    valueEditor: Option[ParameterValueInput],
-    valueCompileTimeValidation: Option[ParameterValueCompileTimeValidation]
+    valueEditor: Option[ParameterValueInput]
 )