+++ designer/client/src/components/graph/node-modal/fragment-input-definition/item/Item.tsx
@@ -75,10 +75,7 @@ export function Item(props: ItemProps): JSX.Element {
                 />
                 <SettingsButton isOpen={isOpen} toggleIsOpen={openSettingMenu} />
             </FieldsRow>
-            <SettingsProvider
-                initialFixedValuesList={item?.valueEditor?.fixedValuesList}
-                initialTemporaryValueCompileTimeValidation={item?.valueCompileTimeValidation}
-            >
+            <SettingsProvider initialFixedValuesList={item?.valueEditor?.fixedValuesList}>
                 {isOpen && (
                     <Settings
                         path={path}