+++ designer/client/src/components/graph/node-modal/fragment-input-definition/settings/SettingsProvider.tsx
@@ -1,32 +1,21 @@
 import React, { createContext, PropsWithChildren, useContext, useState } from "react";
-import { FixedValuesOption, ValueCompileTimeValidation } from "../item";
+import { FixedValuesOption } from "../item";
 
 const SettingsContext = createContext<{
     temporaryUserDefinedList: FixedValuesOption[];
     handleTemporaryUserDefinedList: (listItem: FixedValuesOption[]) => void;
-    temporaryValueCompileTimeValidation: ValueCompileTimeValidation;
-    handleTemporaryValueCompileTimeValidation: (valueCompileTimeValidation: ValueCompileTimeValidation) => void;
 }>(null);
 
 export const SettingsProvider = ({
     initialFixedValuesList,
-    initialTemporaryValueCompileTimeValidation,
     children,
 }: PropsWithChildren<{
     initialFixedValuesList: FixedValuesOption[];
-    initialTemporaryValueCompileTimeValidation: ValueCompileTimeValidation;
 }>) => {
     const [temporaryUserDefinedList, setTemporaryUserDefinedList] = useState<FixedValuesOption[]>(initialFixedValuesList || []);
-    const [temporaryValueCompileTimeValidation, setTemporaryValueCompileTimeValidation] = useState<ValueCompileTimeValidation>(
-        initialTemporaryValueCompileTimeValidation || null,
-    );
 
     const handleTemporaryUserDefinedList = (listItem: FixedValuesOption[]) => {
         setTemporaryUserDefinedList(listItem);
-    };
-
-    const handleTemporaryValueCompileTimeValidation = (valueCompileTimeValidation: ValueCompileTimeValidation) => {
-        setTemporaryValueCompileTimeValidation(valueCompileTimeValidation);
     };
 
     return (
@@ -34,8 +23,6 @@ export const SettingsProvider = ({
             value={{
                 temporaryUserDefinedList,
                 handleTemporaryUserDefinedList,
-                temporaryValueCompileTimeValidation,
-                handleTemporaryValueCompileTimeValidation,
             }}
         >
             {children}