+++ designer/client/test/Process/FrgamentInputDefinition/Item-test.tsx
@@ -36,7 +36,6 @@ describe(Item.name, () => {
                             hintText: "",
                             initialValue: null,
                             valueEditor: null,
-                            valueCompileTimeValidation: null,
                             typ: {
                                 refClazzName: "java.lang.String",
                             } as ReturnedType,