+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/definition/fragment/FragmentParametersDefinitionExtractor.scala
@@ -95,11 +95,7 @@ class FragmentParametersDefinitionExtractor(
       )
       .getOrElse((EditorExtractor.extract(parameterData, ParameterConfig.empty), List.empty))
 
-    val validationExpressionValidator = fragmentParameter.valueCompileTimeValidation.map(validation =>
-      ValidationExpressionParameterValidatorToCompile(validation)
-    )
-
-    val validators = validationExpressionValidator ++ ValidatorsExtractor
+    val validators = ValidatorsExtractor
       .extract(
         ValidatorExtractorParameters(
           ParameterData(typ, Nil),